Split EnvelopeSerialization to separate concerns between publish and consume pipelines
